+++ bindings/angular2/src/directives/ons-tabbar.ts
@@ -1,17 +1,17 @@
 import {
   Component,
-  DynamicComponentLoader,
+  ComponentRef,
+  ViewContainerRef,
+  ComponentResolver,
   Injector,
   Directive,
   ElementRef,
   Type,
-  provide,
-  NgZone,
-  Renderer,
-  Input
+  Input,
+  OnDestroy
 } from '@angular/core';
 
-interface TabbarElement {}
+declare var ons: any;
 
 /**
  * @element ons-tabbar
@@ -24,15 +24,6 @@ interface TabbarElement {}
   selector: 'ons-tabbar'
 })
 export class OnsTabbar {
-  private _tabbar: TabbarElement;
-
-  constructor(private _elementRef: ElementRef) {
-    this._tabbar = _elementRef.nativeElement;
-  }
-
-  get element(): any {
-    return this._tabbar;
-  }
 }
 
 /**
@@ -45,14 +36,44 @@ export class OnsTabbar {
 @Directive({
   selector: 'ons-tab'
 })
-export class OnsTab {
-  private _tab: TabbarElement;
+export class OnsTab implements OnDestroy {
+  private _pageComponent: ComponentRef<any> = null;
 
-  constructor(private _elementRef: ElementRef) {
-    this._tab = _elementRef.nativeElement;
+  constructor(private _elementRef: ElementRef,
+    private _viewContainer: ViewContainerRef,
+    private _resolver: ComponentResolver) {
+
+    // set up ons-tab's page loader
+    this._elementRef.nativeElement.pageLoader = new ons.PageLoader((page, parent, done) => {
+      this._resolver.resolveComponent(page).then(factory => {
+        const pageComponentRef = this._viewContainer.createComponent(factory, 0);
+
+        if (this._pageComponent) {
+          this._pageComponent.destroy();
+        }
+        this._pageComponent = pageComponentRef;
+
+        const pageElement = pageComponentRef.location.nativeElement;
+        parent.appendChild(pageElement); // dirty fix to insert in correct position
+
+        done({
+          element: pageElement,
+          unload: () => {
+            pageComponentRef.destroy();
+          }
+        });
+      });
+    });
   }
 
-  get element(): any {
-    return this._tab;
+  @Input('page') set page(pageComponentType: Type) {
+    this._elementRef.nativeElement.page = pageComponentType;
+  }
+
+  ngOnDestroy() {
+    if (this._pageComponent) {
+      this._pageComponent.destroy();
+    }
   }
 }
+