+++ bindings/angular2/src/ons/notification.ts
@@ -1,1 +1,3 @@
+declare var ons: any;
+
 export const onsNotification = ons.notification;