+++ bindings/angular2/src/ons/platform.ts
@@ -1,2 +1,3 @@
+declare var ons: any;
 
 export const onsPlatform = ons.platform;