+++ core/src/ons/ons.js
@@ -28,6 +28,7 @@ import animationOptionsParser from './animation-options-parser';
 import autoStyle from './autostyle';
 import DoorLock from './doorlock';
 import contentReady from './content-ready';
+import {defaultPageLoader, PageLoader} from './page-loader';
 
 /**
  * @object ons
@@ -51,6 +52,8 @@ ons._animationOptionsParser = animationOptionsParser;
 ons._autoStyle = autoStyle;
 ons._DoorLock = DoorLock;
 ons._contentReady = contentReady;
+ons.defaultPageLoader = defaultPageLoader;
+ons.PageLoader = PageLoader;
 
 ons._readyLock = new DoorLock();