+++ CHANGELOG.md
@@ -4,6 +4,8 @@ CHANGELOG
 
 v2.0.0-rc.16
 ----
+ * ons-tabbar, ons-tab: Added "page" and "pageLoader" property. Changed to use page-loader instead of `ons._internal.getPageHTMLAsync()`.
+ * core: Added ons.defaultPageLoader and ons.PageLoader.
  * ons-page: Changed to accept ".content" and ".background" elements as child contents.
  * ons-modal: Fix [#1433](https://github.com/OnsenUI/OnsenUI/issues/1433).
  * ons-navigator: Improved iOS slide animation (again).
@@ -11,7 +13,6 @@ v2.0.0-rc.16
  * ons-pull-hook, ons-carousel: Fix [#1004](https://github.com/OnsenUI/OnsenUI/issues/1004).
  * ons-fab: Fix [#1496](https://github.com/OnsenUI/OnsenUI/issues/1496).
  * ons-list-item: Fix [#1499](https://github.com/OnsenUI/OnsenUI/issues/1499)
-
 
 v2.0.0-rc.15
 ----